+++ delete_user.h
@@ -1,0 +1,9 @@
+#ifndef DELETE_USER_H
+#define DELETE_USER_H
+
+#include <sqlite3.h>
+
+// ƒ†[ƒU[íœ
+int delete_user(sqlite3* db, int user_id);
+
+#endif
