+++ list_users.h
@@ -1,0 +1,9 @@
+#ifndef LIST_USERS_H
+#define LIST_USERS_H
+
+#include <sqlite3.h>
+
+// ƒ†[ƒU[ˆê——
+void list_users(sqlite3* db);
+
+#endif
