+++ register_user.c
@@ -1,0 +1,48 @@
+#include <stdio.h>
+#include <string.h>
+#include <sqlite3.h>
+#include "register_user.h"
+
+
+#define MAX_LEN 128
+
+// 新規登録
+int register_user(sqlite3* db) {
+    char user[MAX_LEN], pass[MAX_LEN], mail[MAX_LEN];
+    printf("新規ユーザー名: ");
+    if (scanf("%127s", user) != 1) {
+        fprintf(stderr, "入力エラー: ユーザー名が不正です。\n");
+        return -1;
+    }
+    printf("パスワード: ");
+    if (scanf("%127s", pass) != 1) {
+        fprintf(stderr, "入力エラー: パスワードが不正です。\n");
+        return -1;
+    }
+    printf("メールアドレス: ");
+    if (scanf("%127s", mail) != 1) {
+        fprintf(stderr, "入力エラー: メールアドレスが不正です。\n");
+        return -1;
+    }
+
+    const char* sql = "INSERT INTO users(username,password,email) VALUES(?,?,?);";
+    sqlite3_stmt* stmt;
+    int rc = sqlite3_prepare_v2(db, sql, -1, &stmt, NULL);
+    if (rc != SQLITE_OK) {
+        fprintf(stderr, "登録準備エラー: %s\n", sqlite3_errmsg(db));
+        return -1;
+    }
+    sqlite3_bind_text(stmt, 1, user, -1, SQLITE_STATIC);
+    sqlite3_bind_text(stmt, 2, pass, -1, SQLITE_STATIC);
+    sqlite3_bind_text(stmt, 3, mail, -1, SQLITE_STATIC);
+
+    rc = sqlite3_step(stmt);
+    if (rc != SQLITE_DONE) {
+        fprintf(stderr, "登録失敗: %s\n", sqlite3_errmsg(db));
+        sqlite3_finalize(stmt);
+        return -1;
+    }
+    sqlite3_finalize(stmt);
+    printf("ユーザー「%s」を登録しました。\n", user);
+    return 0;
+}
