+++ register_user.h
@@ -1,0 +1,9 @@
+#ifndef REGISTER_USER_H
+#define REGISTER_USER_H
+
+#include <sqlite3.h>
+
+// ÉÜÅ[ÉUÅ[ìoò^
+int register_user(sqlite3* db);
+
+#endif
