--- ui.h
@@ -1,9 +1,0 @@
-//#ifndef UI_H
-//#define UI_H
-//
-//// メニュー表示／入力受付
-//int show_menu(void);
-//
-//int show_register(void);
-//
-//#endif
